refs #47 refactor tests
